@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use App\SMS;
 use Illuminate\Http\Request;
 
 class SMSController extends Controller
@@ -13,16 +12,6 @@ class SMSController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
     {
         //
     }
@@ -41,21 +30,10 @@ class SMSController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\SMS  $sMS
+     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(SMS $sMS)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\SMS  $sMS
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(SMS $sMS)
+    public function show($id)
     {
         //
     }
@@ -64,10 +42,10 @@ class SMSController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\SMS  $sMS
+     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, SMS $sMS)
+    public function update(Request $request, $id)
     {
         //
     }
@@ -75,10 +53,10 @@ class SMSController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\SMS  $sMS
+     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(SMS $sMS)
+    public function destroy($id)
     {
         //
     }
