@@ -6,5 +6,8 @@ use Illuminate\Database\Eloquent\Model;
 
 class SmsSchool extends Model
 {
-    //
+    public function smss()
+    {
+        return $this->hasMany('\App\SMS');
+    }
 }
