@@ -5,11 +5,11 @@ namespace App;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
+use Laravel\Passport\HasApiTokens;
 
 class User extends Authenticatable
 {
-    use Notifiable;
-
+    use Notifiable, HasApiTokens, Notifiable;
     /**
      * The attributes that are mass assignable.
      *
@@ -36,4 +36,43 @@ class User extends Authenticatable
     protected $casts = [
         'email_verified_at' => 'datetime',
     ];
+
+    public function roles()
+    {
+        return $this->belongsToMany('\App\Role','user_role','user_id','role_id');
+    }
+
+    public function hasAnyRole($roles)
+    {
+        if(is_array($roles))
+        {
+            foreach ($roles as $role)
+            {
+                if($this->hasRole($role))
+                {
+                    return true;
+                }
+            }
+        }
+        else if ($this->hasRole($roles))
+        {
+            return true;
+        }
+        return false;
+    }
+
+    public function hasRole($role)
+    {
+        if($this->roles()->where('name',$role)->doesntExist())
+        {
+            return false;
+        }
+        return true;
+    }
+
+    public function userRole($roles)
+    {
+        return $this->roles()->where('name',$roles);
+    }
+
 }
